Remove metrics collector, add metrics to router
