--- modules/metrics-collector/ops/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "metrics-collector",
-  "description": "This is only used to install & pin versions for a few global packages in the prod-mode docker container",
-  "dependencies": {
-    "pino-pretty": "4.3.0"
-  }
-}
