--- modules/metrics-collector/ops/webpack.config.js
@@ -1,61 +1,0 @@
-const CopyPlugin = require("copy-webpack-plugin");
-const path = require("path");
-
-module.exports = {
-  mode: "development",
-  target: "node",
-
-  context: path.join(__dirname, ".."),
-
-  entry: path.join(__dirname, "../src/index.ts"),
-
-  externals: {
-    "pg-native": "commonjs2 pg-native",
-    sqlite3: "commonjs2 sqlite3",
-  },
-
-  node: {
-    __filename: false,
-    __dirname: false,
-  },
-
-  resolve: {
-    mainFields: ["main", "module"],
-    extensions: [".js", ".ts", ".json"],
-    symlinks: false,
-  },
-
-  output: {
-    path: path.join(__dirname, "../dist"),
-    filename: "bundle.js",
-  },
-
-  module: {
-    rules: [
-      {
-        test: /\.js$/,
-        exclude: /node_modules/,
-        use: {
-          loader: "babel-loader",
-          options: {
-            presets: ["@babel/env"],
-          },
-        },
-      },
-      {
-        test: /\.ts$/,
-        exclude: /node_modules/,
-        use: {
-          loader: "ts-loader",
-          options: {
-            configFile: path.join(__dirname, "../tsconfig.json"),
-          },
-        },
-      },
-    ],
-  },
-
-  plugins: [],
-
-  stats: { warnings: false },
-};
