--- modules/metrics-collector/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "@connext/metrics-collector",
-  "version": "0.0.1",
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "build": "rm -rf dist && tsc",
-    "build-bundle": "webpack --config ops/webpack.config.js",
-    "start": "node dist/index.js",
-    "test": "bash ops/test.sh"
-  },
-  "dependencies": {
-    "@ethersproject/contracts": "5.0.8",
-    "@ethersproject/constants": "5.0.7",
-    "@ethersproject/providers": "5.0.15",
-    "@ethersproject/wallet": "5.0.9",
-    "@ethersproject/units": "5.0.8",
-    "@connext/vector-contracts": "0.1.8-beta.6",
-    "@connext/vector-types": "0.1.8-beta.6",
-    "@connext/vector-utils": "0.1.8-beta.6",
-    "@sinclair/typebox": "0.12.7",
-    "core-js": "3.8.1",
-    "fastify": "3.8.0",
-    "fastify-cors": "4.1.0",
-    "fastify-oas": "3.0.8",
-    "pino": "6.7.0",
-    "prom-client": "13.1.0",
-    "ts-nats": "1.2.15"
-  },
-  "devDependencies": {
-    "@types/chai": "4.2.14",
-    "@types/chai-as-promised": "7.1.3",
-    "@types/mocha": "8.0.3",
-    "@types/node": "14.14.6",
-    "babel-loader": "8.1.0",
-    "chai": "4.2.0",
-    "chai-as-promised": "7.1.1",
-    "copy-webpack-plugin": "6.2.1",
-    "mocha": "8.2.0",
-    "nodemon": "2.0.6",
-    "nyc": "15.1.0",
-    "pino-pretty": "4.3.0",
-    "ts-loader": "8.0.7",
-    "ts-mocha": "8.0.0",
-    "ts-node": "9.0.0",
-    "typescript": "4.0.5",
-    "webpack": "4.44.2",
-    "webpack-cli": "4.1.0"
-  }
-}
