+++ modules/router/package.json
@@ -20,6 +20,7 @@
     "@ethersproject/address": "5.0.8",
     "@ethersproject/bignumber": "5.0.12",
     "@ethersproject/constants": "5.0.7",
+    "@ethersproject/contracts": "5.0.8",
     "@ethersproject/providers": "5.0.15",
     "@ethersproject/units": "5.0.8",
     "@sinclair/typebox": "0.12.7",
