+++ modules/router/src/index.ts
@@ -11,6 +11,7 @@ import {
   RestServerNodeService,
   ChannelSigner,
   getPublicIdentifierFromPublicKey,
+  getSignerAddressFromPublicIdentifier,
 } from "@connext/vector-utils";
 import {
   IsAlivePayload,
@@ -23,14 +24,20 @@ import {
   WithdrawalCreatedPayload,
   WithdrawalReconciledPayload,
   WithdrawalResolvedPayload,
+  HydratedProviders,
+  ERC20Abi,
 } from "@connext/vector-types";
-import { Registry } from "prom-client";
+import { Gauge, Registry } from "prom-client";
 import { Wallet } from "ethers";
 
 import { config } from "./config";
 import { IRouter, Router } from "./router";
 import { PrismaStore } from "./services/store";
 import { NatsRouterMessagingService } from "./services/messaging";
+import { AddressZero } from "@ethersproject/constants";
+import { formatEther, formatUnits } from "@ethersproject/units";
+import { Contract } from "@ethersproject/contracts";
+import { BigNumber } from "@ethersproject/bignumber";
 
 const routerPort = 8000;
 const routerBase = `http://router:${routerPort}`;
@@ -84,6 +91,7 @@ const evts: EventCallbackConfig = {
 };
 
 const configuredIdentifier = getPublicIdentifierFromPublicKey(Wallet.fromMnemonic(config.mnemonic).publicKey);
+const configuredSigner = getSignerAddressFromPublicIdentifier(configuredIdentifier);
 
 const logger = pino({ name: configuredIdentifier });
 logger.info({ config }, "Loaded config from environment");
@@ -94,6 +102,54 @@ server.register(metricsPlugin, { endpoint: "/metrics", prefix: "router_" });
 
 let router: IRouter;
 const store = new PrismaStore();
+
+const hydrated: HydratedProviders = hydrateProviders(config.chainProviders);
+
+// create gauge for each rebalanced asset and each native asset for the signer address
+// TODO: maybe want to look into an API rather than blowing up our eth providers? although its not that many calls
+const gauges: { [chainId: string]: { [asset: string]: Gauge<string> } } = {};
+Object.entries(hydrated).forEach(([chainId, provider]) => {
+  // base asset
+  gauges[chainId] = {};
+  gauges[chainId][AddressZero] = new Gauge({
+    name: `chain_${chainId}_base_asset`,
+    help: `chain_${chainId}_base_asset_help`,
+    registers: [register],
+    async collect() {
+      const balance = await provider.getBalance(configuredSigner);
+      // NOTE: seems to not be getting scraped at interval
+      console.log("***** calling set on base gauge");
+      this.set(parseFloat(formatEther(balance)));
+    },
+  });
+
+  // get all non-zero addresses
+  const assets = config.rebalanceProfiles
+    .filter((prof) => prof.chainId.toString() === chainId && prof.assetId !== AddressZero)
+    .map((p) => p.assetId);
+
+  const tokens: { [asset: string]: { contract: Contract; decimals?: BigNumber } } = {};
+  assets.forEach((asset) => {
+    tokens[asset] = {
+      contract: new Contract(asset, ERC20Abi, provider),
+      decimals: undefined,
+    };
+  });
+  assets.forEach((assetId) => {
+    gauges[chainId][assetId] = new Gauge({
+      name: `chain_${chainId}_asset_${assetId}`,
+      help: `chain_${chainId}_asset_${assetId}_help`,
+      registers: [register],
+      async collect() {
+        const decimals = tokens[assetId].decimals ?? (await tokens[assetId].contract.functions.decimals());
+        const balance = await tokens[assetId].contract.balanceOf(configuredSigner);
+        // NOTE: seems to not be getting scraped at interval
+        console.log("***** calling set on asset gauge");
+        this.set(parseFloat(formatUnits(balance, decimals)));
+      },
+    });
+  });
+});
 
 server.addHook("onReady", async () => {
   const signer = new ChannelSigner(Wallet.fromMnemonic(config.mnemonic).privateKey);
